requirements.txt
Flask
feedparser
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
